refactor: Added documentation for loading methods
